Add table name for page type class
SilverStripe 4 will use the fully qualified class name by default
when building the database. This is not ideal, and it is recommended
to set a table_name config property. Although this class will not
create a table by default, we do not know if someone might have
applied extensions that would. Thus, set table name for good practices.